changed IDs to Integer instead of Long for simplicity of example

+++ demo/src/main/java/com/senacor/bitc/demo/domain/CustomerRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface CustomerRepository extends CrudRepository<Customer, Long> {
+public interface CustomerRepository extends CrudRepository<Customer, Integer> {
 
     List<Customer> findByLastName(String lastName);
